deletes file
